updated reset to clear out teleopSpeaker, updated tunable number default throughout code

+++ src/main/java/frc/robot/autos/AutoCommandBuilder.java
@@ -85,7 +85,7 @@ public class AutoCommandBuilder {
     return this;
   }
 
-  public AutoCommandBuilder shoot(boolean snapToSpeaker, int setpoint) {
+  public AutoCommandBuilder shoot(boolean snapToSpeaker, double setpoint) {
     if (snapToSpeaker) {
       commands.addCommands(
           new SnapToTargetCommand(
